Add pyproject.toml using poetry
[Poetry][] is a tool widely used in the Python community to help
develop, build, package, and publish Python projects.

With this boilerplate you can use for example `poetry build` to create
wheels and sdists. See the Poetry docs for more options.

Popular alternative tools for packaging++ are setuptools and
flit, which you can find listed with a lot of other projects for Python
packaging on <https://packaging.python.org/key_projects/>. I choose
Poetry as it's proven to be a reasonable (and popular) default, not
because of any analysis for our specific use case.

[Poetry]: https://python-poetry.org