Refactor test connections (#257)
Add test command line options:

  --address=ADDRESS     the Mesh server address. Default localhost:50051.
  --creds-type={insecure,tls,kerberos}
                        the connection credential type. Default insecure.
  --tls-root-certs=FILE
                        use this PEM formatted root certificate chain file.
  --kerberos-service-principal=PRINCIPAL
                        use this service principal name for Kerberos authentication.
  --kerberos-user-principal=PRINCIPAL
                        use this client principal for Kerberos authentication.

These command line options are only available when running pytest in the
src/volue/mesh/tests directory or with that directory as an argument.
A future commit will move tests to /tests/, and then command line
options will work when running from root.

---

Remove the port command line option from examples. The options are now
`example.py ADDRESS TLS_ROOT_CERTS` where `ADDRESS` is on the form
`localhost:50051` and `TLS_ROOT_CERTS` is the path to a certificate
file.

This allows more address formats, see [gRPC name resolution][1],
and is more consistent with the new constructors.

[1]: https://grpc.github.io/grpc/core/md_doc_naming.html,

---

Remove ServerConfig and add test fixtures for connection and session
generation and use those fixtures in all tests.

For synchronous tests the same connection will be reused accross all
tests except authentication tests. For asynchronous tests the connection
will be recreated every test. This is because pytest by default
recreates the async event loop every test, it is fully possible to
change that and use the same event loop for all tests, but that is out
of scope for this PR.

---

- Use new Connection constructors (.insecure, .with_tls, .with_kerberos)
  in tests. Examples should also be updated with these, but this commit
  is large enough as it is.

- Parameterize the tests that run examples on each examples instead of
  running all examples in one tests. This makes it much more clear from
  test failures which example is failing.

- We had two tests named `test_open_and_close_session`, rename one to
  `test_async_open_and_close_session` so we run both.

- Update `if __name__ == '__main__':` blocks in tests to run
  `sys.exit(pytest.main(sys.argv))` instead of `pytest.main()`, this
  will allow passing arguments when explicitly running tests, and result
  in non-zero exit codes when the tests fail.