Use `CaseGroup/CaseName` to refer to hydsim cases
`CaseGroup/CaseName` is established syntax used in various UIs, Optimal
Gateway, and Automation Framework. As discussed in
https://github.com/Volue/energy-sim/issues/742#issuecomment-1811942095.

There's an argument for more user friendly error handling here, but I
think this is API compatible with a future improvement, and I want to
get this API break done ASAP. For now you'll get a `ValueError` if the
split doesn't split.

Closes https://github.com/Volue-Public/energy-mesh-python/issues/398.