Add support for getting HydSim datasets (#453)
A HydSim dataset is a representation of data used in a hydro simulation
or inflow calculation. The HydSim team sometimes uses datasets when
diagnosing customer issues. Traditionally datasets have been stored on
the Mesh server if an attribute with a directory path is non-empty in
the Mesh model.

A problem with this setup is that most users shouldn't have access to
the Mesh server. When a regular user wants to report a HydSim issue
they may need help from a superuser to extract the datasets from the
Mesh server. This is not ideal.

Going forward we want Mesh clients that run simulations and inflow
calculations to handle datasets on the client side so that it's
possible for regular users to access datasets, as well as more
convenient for users that might have Mesh access.

This patch adds support for getting HydSim datasets when requesting
simulations and inflow calculations using Python. If requested the
datasets are streamed to the client together with the log messages we
already stream.

See https://github.com/Volue/energy-mesh/pull/5260 (internal) for the
server implementation.